+++ CHANGELOG.md
@@ -4,6 +4,10 @@ The changelog for [ApplozicSwift](https://github.com/AppLozic/ApplozicSwift). Al
 
 2.6.1
 ---
+
+### Enhancments
+- [AL-3391] Add support for autosuggestions.
+
 ### Fixes
 - [AL-3533] Fixed an issue where scrolling in group detail screen would cause inconsistency in profile images of participants.
 - [AL-3533] Fixed an issue where app was crashing while removing table viewcontroller.